cleaned up files and merged date utils into single place
